day3-network-bridge/python-flask-mysql/requirements.txt
Flask==0.12.2
mysql-connector-python
